restructure properties module
